# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAMBINA",
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://sabina.dev:/api"
        }
    ],
    "paths": {
...
        "/user/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "User Me",
                "description": "Get Me Data",
                "operationId": "api.user.index",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully get data"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": "1"
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Administrator"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            }
        },
        "/user/store": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "User Registration",
                "operationId": "api.user.store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "ktp",
                                    "password",
                                    "password_confirmation",
                                    "placeofbirth",
                                    "phone_number",
                                    "address",
                                    "gender",
                                    "dateofbirth"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "jhondoe@gmail.com"
                                    },
                                    "password": {
                                        "type": "password",
                                        "example": "Qwer1234!"
                                    },
                                    "password_confirmation": {
                                        "type": "password",
                                        "example": "Qwer1234!"
                                    },
                                    "placeofbirth": {
                                        "type": "text",
                                        "example": "Jakarta"
                                    },
                                    "dateofbirth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1991-02-24"
                                    },
                                    "address": {
                                        "type": "text",
                                        "example": "Jl Bunga No 12"
                                    },
                                    "gender": {
                                        "description": "1 lanang 2 wadon",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "phone_number": {
                                        "type": "text",
                                        "example": "08123456789"
                                    },
                                    "ktp": {
                                        "type": "file"
                                    },
                                    "avatar": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully save data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "mobilekey": []
                    }
                ]
            }
        },
        "/user/update-my-profile": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update My Profile",
                "operationId": "api.user.updatemyprofile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "placeofbirth",
                                    "phone_number",
                                    "address",
                                    "gender",
                                    "dateofbirth"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "text",
                                        "example": "John"
                                    },
                                    "email": {
                                        "type": "text",
                                        "example": "jhondoe@gmail.com"
                                    },
                                    "password": {
                                        "description": "Biarkan kosong jika tidak diganti",
                                        "type": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "Biarkan kosong jika tidak diganti",
                                        "type": "password"
                                    },
                                    "placeofbirth": {
                                        "type": "text",
                                        "example": "Jakarta"
                                    },
                                    "dateofbirth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1991-02-24"
                                    },
                                    "address": {
                                        "type": "text",
                                        "example": "Jl Bunga No 12"
                                    },
                                    "gender": {
                                        "description": "1 lanang 2 wadon",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "phone_number": {
                                        "type": "text",
                                        "example": "08123456789"
                                    },
                                    "ktp": {
                                        "type": "file"
                                    },
                                    "avatar": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Berhasil memperbarui data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            }
        },
        "/user/my-room": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get My Room",
                "description": "Get My Room",
                "operationId": "api.user.myroom",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully get data"
                                        },
                                        "data": {
                                            "properties": {
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "room_name": {
                                                                "type": "integer",
                                                                "example": "room name"
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "room blablabla"
                                                            },
                                                            "entry_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2023-02-24"
                                                            },
                                                            "checkout_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2023-03-03"
                                                            },
                                                            "image_path": {
                                                                "type": "string",
                                                                "example": "path image"
                                                            },
                                                            "it_will_end": {
                                                                "type": "string",
                                                                "example": "1 = will end , 0 = not yet"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": [],
                        "mobilekey": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "mobilekey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    }
}